+++ test/mocks.c
@@ -113,7 +113,7 @@ unpatch_all(void)
 
 int
 dma_controller_add_region(dma_controller_t *dma, void *dma_addr,
-                          size_t size, int fd, off_t offset,
+                          uint64_t size, int fd, off_t offset,
                           uint32_t prot)
 {
     if (!is_patched("dma_controller_add_region")) {
